Fix #125 and small code refactoring
