+++ omnimatter/prototypes/buildings/omnifurnace.lua
@@ -1,0 +1,134 @@
+BuildGen:create("omnimatter","omni-furnace"):
+	setEnergy(5):
+	setIngredients({ "omnicium-plate", 5},{ "omnite-brick", 5},{ "stone-furnace", 1}):
+	setStacksize(20):
+	setSubgroup("omni-buildings"):
+	setSize(2):
+	setCrafting("smelting","omnifurnace"):
+	setUsage(45):
+	setEnabled():
+	setBurner(1,1):
+	setAnimation({
+      layers = {
+        {
+          filename = "__omnimatter__/graphics/entity/buildings/omni-furnace.png",
+          priority = "high",
+          width = 85,
+          height = 87,
+          frame_count = 1,
+          shift = util.by_pixel(-1.5, 1.5),
+          hr_version = {
+            filename = "__omnimatter__/graphics/entity/buildings/hr-omni-furnace.png",
+            priority = "high",
+            width = 171,
+            height = 174,
+            frame_count = 1,
+            shift = util.by_pixel(-1.25, 2),
+            scale = 0.5
+          }
+        },
+        {
+          filename = "__base__/graphics/entity/steel-furnace/steel-furnace-shadow.png",
+          priority = "high",
+          width = 139,
+          height = 43,
+          frame_count = 1,
+          draw_as_shadow = true,
+          shift = util.by_pixel(39.5, 11.5),
+          hr_version = {
+            filename = "__base__/graphics/entity/steel-furnace/hr-steel-furnace-shadow.png",
+            priority = "high",
+            width = 277,
+            height = 85,
+            frame_count = 1,
+            draw_as_shadow = true,
+            shift = util.by_pixel(39.25, 11.25),
+            scale = 0.5
+          }
+        },
+      },
+    }):
+	setWorkVis({
+      {
+        north_position = {0.0, 0.0},
+        east_position = {0.0, 0.0},
+        south_position = {0.0, 0.0},
+        west_position = {0.0, 0.0},
+        animation =
+        {
+          filename = "__base__/graphics/entity/steel-furnace/steel-furnace-fire.png",
+          priority = "high",
+          line_length = 8,
+          width = 29,
+          height = 40,
+          frame_count = 48,
+          axially_symmetrical = false,
+          direction_count = 1,
+          shift = util.by_pixel(-0.5, 6),
+          hr_version = {
+            filename = "__base__/graphics/entity/steel-furnace/hr-steel-furnace-fire.png",
+            priority = "high",
+            line_length = 8,
+            width = 57,
+            height = 81,
+            frame_count = 48,
+            axially_symmetrical = false,
+            direction_count = 1,
+            shift = util.by_pixel(-0.75, 5.75),
+            scale = 0.5
+          }
+        },
+        light = {intensity = 1, size = 1, color = {r = 1.0, g = 1.0, b = 1.0}}
+      },
+      {
+        north_position = {0.0, 0.0},
+        east_position = {0.0, 0.0},
+        south_position = {0.0, 0.0},
+        west_position = {0.0, 0.0},
+        effect = "flicker", -- changes alpha based on energy source light intensity
+        animation =
+        {
+          filename = "__base__/graphics/entity/steel-furnace/steel-furnace-glow.png",
+          priority = "high",
+          width = 60,
+          height = 43,
+          frame_count = 1,
+          shift = {0.03125, 0.640625},
+          blend_mode = "additive"
+        }
+      },
+      {
+        north_position = {0.0, 0.0},
+        east_position = {0.0, 0.0},
+        south_position = {0.0, 0.0},
+        west_position = {0.0, 0.0},
+        effect = "flicker", -- changes alpha based on energy source light intensity
+        animation =
+        {
+          filename = "__base__/graphics/entity/steel-furnace/steel-furnace-working.png",
+          priority = "high",
+          line_length = 8,
+          width = 64,
+          height = 75,
+          frame_count = 1,
+          axially_symmetrical = false,
+          direction_count = 1,
+          shift = util.by_pixel(0, -4.5),
+          blend_mode = "additive",
+          hr_version = {
+            filename = "__base__/graphics/entity/steel-furnace/hr-steel-furnace-working.png",
+            priority = "high",
+            line_length = 8,
+            width = 130,
+            height = 149,
+            frame_count = 1,
+            axially_symmetrical = false,
+            direction_count = 1,
+            shift = util.by_pixel(0, -4.25),
+            blend_mode = "additive",
+            scale = 0.5
+          }
+        }
+      },
+    }):
+	setReplace("furnace"):extend()
